% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alfa.R
\name{alfa_c}
\alias{alfa_c}
\alias{alfa_fr}
\title{Szacowanie rzetelnosci}
\usage{
alfa_c(x, na.rm = TRUE, verbose = TRUE)

alfa_fr(x, na.rm = TRUE, verbose = TRUE)
}
\arguments{
\item{x}{macierz typu \code{numeric} lub ramka danych (data frame)
zawierająca zmienne typu \code{numeric}}

\item{na.rm}{wartość logiczna - czy przy obliczeniach ignorować braki danych}

\item{verbose}{wartość logiczna - czy wydrukować wyniki analizy}
}
\value{
Funkcje zwracają milcząco trzyelementową listę, której elementy zawierają:
\itemize{
   \item{\code{alfa} wartość współczynnika alfa Cronbacha/Feldt-Raju dla
         całego testu,}
   \item{\code{alfaBZ} wektor liczbowy zawierający współczynniki alfa
         Cronbacha/Feldt-Raju obliczone dla testu z wyłączeniem danego
         zadania,}
   \item{\code{wsp = ("alfa Cronbacha" | "alfa Feldt-Raju").}}
}
}
\description{
Funkcje służące do oszacowania rzeteleności testu przy pomocy
współczynnika alfa Cronbacha lub alfa Feldt-Rahu.
}
\details{
Funkcja \code{alfa_c} wykorzystuje współczynnik alfa Cronbacha - szerzej
rozpoznawalny i stosowany, ale oparty na modelu nakładającym na test bardziej
restrykcyjne założenia (taka sama siła związku pomiędzy mierzoną cechą,
a wynikiem każdego zadania). W efekcie alfa Cronbacha może nieco zaniżać
oszacowanie rzetelności testu.

Funkcja  \code{alfa_fr} wykorzystuje współczynnik alfa Feldt-Raju,
o analogicznej interpretacji i podobnej konstrukcji, ale dopuszczający różną
siłę związku pomiędzy mierzoną cechą a wynikami każdego zadania. Jego użycie
jest bardziej adekwatne szczególnie w sytuacji, gdy w teście występuje duża
rozpiętość trudności zadań i/lub różne zadania mają różne skale oceny. Jego
wartość jest zawsze nie mniejsza niż wartość współczynnika alfa Cronbacha.
}
\examples{
alfa_c(wynikiSymTest)
alfa_fr(wynikiSymTest)
}
\seealso{
\code{\link{parametry_zadan}}, \code{\link{wykres_tmr}}
}

