% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alfa.R
\name{alfaFR}
\alias{alfaFR}
\title{Szacowanie rzetelności}
\usage{
alfaFR(x, na.rm = TRUE)
}
\arguments{
\item{x}{macierz typu \code{numeric} lub ramka danych (data frame)
zawierająca zmienne typu \code{numeric}}

\item{na.rm}{wartość logiczna - czy przy obliczeniach ignorować braki danych}
}
\value{
Funkcja zwraca milcząco trzyelementową listę, której elementy zawierają:
\itemize{
   \item{\code{alfa} wartość współczynnika alfa Cronbacha dla całego testu,}
   \item{\code{alfaBZ} wektor liczbowy zawierający współczynniki alfa
         Cronbacha obliczone dla testu z wyłączeniem danego zadania,}
   \item{\code{wsp = "alfa Feldt-Raju".}}
}
}
\description{
Funkcja służy do oszacowania rzeteleności testu przy pomocy
współczynnika alfa Feldt-Raju. Zwraca oszacowanie dla całego testu oraz
z wyłączeniem poszczególnych zadań.
}

