% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert_maks}
\alias{assert_maks}
\title{Sprawdzanie argumentów funkcji}
\usage{
assert_maks(maks, x)
}
\arguments{
\item{maks}{argument do sprawdzenia - wektor minimalnych wartości lub NULL}

\item{x}{argument, względem którego \code{x} ma być zweryfikowany - macierz
lub ramka danych}
}
\value{
Funkcja zwraca wektor o liczbie elementóW odpowiadającej liczie kolumn
\code{x}, który zawiera minimalnych możliwych do przyjęcia wartości.
}
\description{
Funkcja sprawdza, czy argument jest pasującym do odpowiedniej
macierzy/ramki danych wektorem opisującym minimalne możliwe do przyjęcia
wartości.
}

