% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert_mdfn}
\alias{assert_mdfn}
\title{Sprawdzanie argumentow funkcji}
\usage{
assert_mdfn(x, nazwaArgumentu = "x")
}
\arguments{
\item{x}{argument do sprawdzenia}

\item{nazwaArgumentu}{nazwa argumentu, która zostanie wstawiona do
komunikatów o błędzie}
}
\value{
Funkcja zwraca \code{TRUE} jeśli argument jest macierzą liczbową lub ramką
daych zawierającą tylko zmienne liczbowe. W przeciwnym wypadku wywołuje błąd.
}
\description{
Funkcja sprawdza, czy argument jest macierz liczbową lub ramką
danych zawierającą tylko zmienne liczbowe.
}

