% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert_mm}
\alias{assert_mm}
\title{Sprawdzanie argumentow funkcji}
\usage{
assert_mm(x, y, nazwaArgumentu = "x")
}
\arguments{
\item{x}{argument do sprawdzenia - wektor}

\item{y}{argument, względem którego \code{x} ma być zweryfikowany - macierz
lub ramka danych}

\item{nazwaArgumentu}{nazwa argumentu, która zostanie wstawiona do
komunikatów o błędzie}
}
\value{
Funkcja zwraca \code{TRUE} jeśli argument \code{x} jest poprawny.
W przeciwnym wypadku wywołuje błąd.
}
\description{
Funkcja sprawdza, czy argument jest wektorem liczb pasującym
do odpowiedniej macierzy/ramki danych.
}

