% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_ot.R
\name{assert_ot}
\alias{assert_ot}
\title{Sprawdzanie argumentów funkcji}
\usage{
assert_ot(x, y, nazwaArgumentuX = "x", nazwaArgumentuY = "opisTestu")
}
\arguments{
\item{x}{argument do sprawdzenia - dane}

\item{y}{argument do sprawdzenia - opis struktury testu}

\item{nazwaArgumentuX}{nazwa argumentu, która zostanie wstawiona do
komunikatów o błędzie}

\item{nazwaArgumentuY}{nazwa argumentu, która zostanie wstawiona do
komunikatów o błędzie}
}
\value{
Funkcja zwraca \code{TRUE} jeśli argument \code{y} jest poprawnym opisem
struktury testu oraz dane podane argumentem \code{x} są z nim zgodne.
W przeciwnym wypadku wywołuje błąd.
}
\description{
Funkcja sprawdza, czy podane dane są zgodne podanym opisem
struktury testu (oraz czy sam ten opis jest poprawny).
}

