% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert_t}
\alias{assert_t}
\title{Sprawdzanie argumentow funkcji}
\usage{
assert_t(x, nazwaArgumentu = "x")
}
\arguments{
\item{x}{argument do sprawdzenia}

\item{nazwaArgumentu}{nazwa argumentu, która zostanie wstawiona do
komunikatów o błędzie}
}
\value{
Funkcja zwraca \code{TRUE} jeśli argument jest ciągiem znaków
(jednoelementowym wektorem typu character). W przeciwnym wypadku wywołuje
błąd.
}
\description{
Funkcja sprawdza, czy argument jest ciągiem znaków.
}

