% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daneEWD.R
\docType{data}
\name{daneEWD}
\alias{daneEWD}
\title{Test wiedzy o EWD - dane}
\format{data frame: 124 obserwacje, 42 zmienne}
\description{
Wyniki testu wiedzy o EWD przeprowadzonego na uczestnikach
\emph{Jesiennej Szkoły EWD 2015}.

\bold{Opis zmiennych w zbiorze:}
\itemize{
  \item{\code{id} identyfikator rozwiązującego}
  \item{\code{grupa} identyfikator grupy, w której uczestniczył rozwiązujący}
  \item{\code{I1} Blok stwierdzeń prawda/fałsz - stwierdzenie 1. [P/\bold{F}]}
  \item{\code{I2} Blok stwierdzeń prawda/fałsz - stwierdzenie 2. [P/\bold{F}]}
  \item{\code{I3} Blok stwierdzeń prawda/fałsz - stwierdzenie 3. [P/\bold{F}]}
  \item{\code{I4} Blok stwierdzeń prawda/fałsz - stwierdzenie 4. [P/\bold{F}]}
  \item{\code{I5} Blok stwierdzeń prawda/fałsz - stwierdzenie 5. [P/\bold{F}]}
  \item{\code{I6} Blok stwierdzeń prawda/fałsz - stwierdzenie 6. [\bold{P}/F]}
  \item{\code{I7} Blok stwierdzeń prawda/fałsz - stwierdzenie 7. [P/\bold{F}]}
  \item{\code{I8} Blok stwierdzeń prawda/fałsz - stwierdzenie 8. [P/\bold{F}]}
  \item{\code{I9} Blok stwierdzeń prawda/fałsz - stwierdzenie 9. [P/\bold{F}]}
  \item{\code{I10} Blok stwierdzeń prawda/fałsz - stwierdzenie 10. [P/\bold{F}]}
  \item{\code{I11} Blok stwierdzeń prawda/fałsz - stwierdzenie 11. [P/\bold{F}]}
  \item{\code{I12} Blok stwierdzeń prawda/fałsz - stwierdzenie 12. [P/\bold{F}]}
  \item{\code{I13} Blok stwierdzeń prawda/fałsz - stwierdzenie 13. [\bold{P}/F]}
  \item{\code{I14} Blok stwierdzeń prawda/fałsz - stwierdzenie 14. [P/\bold{F}]}
  \item{\code{I15} Blok stwierdzeń prawda/fałsz - stwierdzenie 15. [P/\bold{F}]}
  \item{\code{I16} Blok stwierdzeń prawda/fałsz - stwierdzenie 16. [P/\bold{F}]}
  \item{\code{I17} Blok stwierdzeń prawda/fałsz - stwierdzenie 17. [\bold{P}/F]}
  \item{\code{I18} Blok stwierdzeń prawda/fałsz - stwierdzenie 18. [\bold{P}/F]}
  \item{\code{I19} Blok stwierdzeń prawda/fałsz - stwierdzenie 19. [P/\bold{F}]}
  \item{\code{I20} Blok stwierdzeń prawda/fałsz - stwierdzenie 20. [\bold{P}/F]}
  \item{\code{Z1} Pytania wielokrotnych odpowiedzi - pytanie 1. [A/B/\bold{C}/D]}
  \item{\code{Z2} Pytania wielokrotnych odpowiedzi - pytanie 2. [PA/\bold{PB}/FC/FD]}
  \item{\code{Z3} Pytania wielokrotnych odpowiedzi - pytanie 3. [A/B/\bold{C}/D]}
  \item{\code{Z4} Pytania wielokrotnych odpowiedzi - pytanie 4. [A/\bold{B}/C/D]}
  \item{\code{Z5} Pytania wielokrotnych odpowiedzi - pytanie 5. [\bold{A}/B/C/D]}
  \item{\code{Z6} Pytania wielokrotnych odpowiedzi - pytanie 6. [A/\bold{B}/C/D]}
  \item{\code{Z7} Pytania wielokrotnych odpowiedzi - pytanie 7. [\bold{A}/B/C/D]}
  \item{\code{Z8.1} Pytania wielokrotnych odpowiedzi - pytanie 8.1 [\bold{P}/F]}
  \item{\code{Z8.2} Pytania wielokrotnych odpowiedzi - pytanie 8.2. [\bold{P}/F]}
  \item{\code{Z8.3} Pytania wielokrotnych odpowiedzi - pytanie 8.3. [P/\bold{F}]}
  \item{\code{Z8.4} Pytania wielokrotnych odpowiedzi - pytanie 8.4. [\bold{P}/F]}
  \item{\code{Z8.5} Pytania wielokrotnych odpowiedzi - pytanie 8.5. [P/\bold{F}]}
  \item{\code{Z9} Pytania wielokrotnych odpowiedzi - pytanie 9. [A/B/\bold{C}/D]}
  \item{\code{Z10} Pytania wielokrotnych odpowiedzi - pytanie 10. [A/\bold{B}/C/D]}
  \item{\code{Z11} Pytania wielokrotnych odpowiedzi - pytanie 11. [A/B/C/\bold{D}]}
  \item{\code{Z12} Pytania wielokrotnych odpowiedzi - pytanie 12. [A/B/C/\bold{D}]}
  \item{\code{Z13} Pytania wielokrotnych odpowiedzi - pytanie 13. [A/B/\bold{C}/D]}
  \item{\code{Z14.1} Pytania wielokrotnych odpowiedzi - pytanie 14.1. [P/\bold{F}]}
  \item{\code{Z14.2} Pytania wielokrotnych odpowiedzi - pytanie 14.2. [P/\bold{F}]}
  \item{\code{Z14.3} Pytania wielokrotnych odpowiedzi - pytanie 14.3. [\bold{P}/F]}
}
}
\seealso{
\code{\link{opisTestuEWD}}
}
\keyword{datasets}

