% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normy_staninowe.R
\name{grupuj_staniny}
\alias{grupuj_staniny}
\title{Normy staninowe dla sumy punktow}
\usage{
grupuj_staniny(x)
}
\arguments{
\item{x}{wektor liczb całkowitych z przedziału [1;9]}
}
\value{
wektor ciągów tekstów
}
\description{
Funkcja przekodowuje wynik wyrażony w staninach na wartości
opisujące przydział do trzech grup: 1-3, 4-6, 7-9 stanin.
}
\examples{
x = normy_staninowe(wynikiSymTest)$wynikiStaninowe
ftable(x, grupuj_staniny(x))
}

