% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_macierz_danych.R
\name{info_macierz_danych}
\alias{info_macierz_danych}
\title{Informacje o danych}
\usage{
info_macierz_danych(x)
}
\arguments{
\item{x}{macierz ramka danych (data frame)}
}
\value{
Funkcja zwraca milcząco ciąg znaków.
}
\description{
Funkcja zwraca ciąg znaków zawierający podstawowe informacje
o macierzy danych.
}

