% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infoMacierzDanych.R
\name{infoMacierzDanych}
\alias{infoMacierzDanych}
\title{Informacje o danych}
\usage{
infoMacierzDanych(x)
}
\arguments{
\item{x}{macierz ramka danych (data frame)}
}
\value{
Funkcja zwraca milcząco ciąg znaków.
}
\description{
Funkcja zwraca ciąg znaków zawierający podstawowe informacje
o macierzy danych.
}

