% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moc_roznicujaca.R
\name{moc_roznicujaca}
\alias{moc_roznicujaca}
\title{Moc różnicująca}
\usage{
moc_roznicujaca(x, na.rm = TRUE)
}
\arguments{
\item{x}{macierz typu \code{numeric} lub ramka danych (data frame)
zawierająca zmienne typu \code{numeric}}

\item{na.rm}{wartość logiczna - czy przy obliczeniach ignorować braki danych}
}
\value{
Funkcja zwraca milcząco dwuelementową listę, której elementy zawierają:
\itemize{
   \item{\code{korelacjePearsona} wartości współczynników korelacji Pearsona
         (punktowo-dwuseryjnych) pomiędzy wynikiem całego testu, a wynikami
         poszczególnych zadań,}
   \item{\code{korelacjeDwuseryjne} wartości współczynników korelacji
         dwuseryjnych pomiędzy wynikiem całego testu, a wynikami
         poszczególnych zadań.}
}
}
\description{
Funkcja służy do oszacowania mocy różnicującej zadania.
}

