% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opisTestuEWD.R
\docType{data}
\name{opisTestuEWD}
\alias{opisTestuEWD}
\title{Test wiedzy o EWD - opis zadań}
\format{data frame: 42 obserwacje, 6 zmiennych}
\description{
Opis zadań użytych w teście wiedzy o EWD przeprowadzonego na uczestnikach
\emph{Jesiennej Szkoły EWD 2015}.
}
\seealso{
\code{\link{daneEWD}}
}
\keyword{datasets}

