% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opisTestuSpr2005.R
\docType{data}
\name{opisTestuSpr2005}
\alias{opisTestuSpr2005}
\title{Sprawdzian 2005 - opis zadan}
\format{data frame: 40 obserwacje, 6 zmiennych}
\description{
Opis struktury zbioru z wynikami sprawdzianu w VI klasie SP z 2005 r.
}
\seealso{
\code{\link{daneSpr2005}}
}
\keyword{datasets}

