% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opisTestuWzrost.R
\docType{data}
\name{opisTestuWzrost}
\alias{opisTestuWzrost}
\title{Ankieta na temat wzrostu - opis zadan}
\format{data frame: 14 obserwacji, 6 zmiennych}
\description{
Opis zadań użytych w ankiecie dotyczącej wzrostu, przeprowadzonej wśród
uczestnikówtrzeciego \emph{Spotkania Entuzjastów R}, które odbyło się
w Warszawie 24 kwietnia 2014 r. (p. <http://smarterpoland.pl/SER/#SERIII>).
}
\seealso{
\code{\link{daneWzrost}}
}
\keyword{datasets}

