% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametry_zadan.R
\name{parametry_zadan}
\alias{parametry_zadan}
\title{Parametry zadan}
\usage{
parametry_zadan(x, maks = NULL, min = NULL, na.rm = TRUE,
  verbose = TRUE)
}
\arguments{
\item{x}{macierz typu \code{numeric} lub ramka danych (data frame)
zawierająca zmienne typu \code{numeric}}

\item{maks}{opcjonalnie wektor liczb całkowitych opisujący maksymalną
liczbę puntków możliwych do uzyskania za poszczególne zadania}

\item{min}{opcjonalnie wektor liczb całkowitych opisujący minimalną
wartość, jaką może przyjąć wynik poszczególnych zadań}

\item{na.rm}{wartość logiczna - czy przy obliczeniach ignorować braki danych}

\item{verbose}{wartość logiczna - czy wydrukować wyniki analizy}
}
\value{
Funkcja zwraca milcząco listę z parametrami zadań.
}
\description{
Funkcja szacuje parametry zadań: łatwości/trudności, moc
różnicującą, rzetelność testu z wyłączeniem zadania.
}
\examples{
parametry_zadan(wynikiSymTest)
}
\seealso{
\code{\link{latwosc}}, \code{\link{moc_roznicujaca}},
\code{\link{alfa_c}}, \code{\link{wykres_tmr}}
}

