% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raport.R
\name{raport}
\alias{raport}
\title{Generowanie raportu}
\usage{
raport(x, tytul = "Tytul raportu", data = NULL, maks = NULL,
  na.rm = TRUE)
}
\arguments{
\item{x}{macierz typu \code{numeric} lub ramka danych (data frame)
zawierająca zmienne typu \code{numeric}}

\item{tytul}{tytuł raportu (ciąg znaków)}

\item{data}{opcjonalnie data (ciąg znaków)}

\item{maks}{opcjonalnie wektor liczb całkowitych opisujący maksymalną
liczbę puntków możliwych do uzyskania za poszczególne zadania}

\item{na.rm}{wartość logiczna - czy przy obliczeniach ignorować braki danych}
}
\value{
Funkcja nic nie zwraca.
}
\description{
Funkcja generuje raport zawierający wyniki analizy
psyhometrycznej testu.
}
\examples{
raport(wynikiSymTest)
}
\seealso{
\code{\link{parametry_zadan}}, \code{\link{wykres_lmr}}
}

