% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSymTest.R
\docType{data}
\name{symTest}
\alias{symTest}
\title{Symulowane wyniki testu złożonego z 10 zadań.}
\format{data frame: 500 obs. of 10 variables}
\description{
Symulowane wyniki testu złożonego z 10 zadań.
}
\keyword{datasets}

