% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wykres_lmr.R
\name{wykres_lmr}
\alias{wykres_lmr}
\title{Parametry zadań}
\usage{
wykres_lmr(x, wsk = "Pearson")
}
\arguments{
\item{x}{macierz typu \code{numeric} lub ramka danych (data frame)
zawierająca zmienne typu \code{numeric}}

\item{wsk}{wskaźnik mocy różnicującej, który ma być użyty: 'Pearson',
'dwuseryjna' lub 'bez zadania'}
}
\value{
Funkcja nic nie zwraca.
}
\description{
Funkcja rysuje wykres trudności i mocy różnicującej zadań.
}
\details{
\bold{Interpretacja wykresu}
Ogólnie rzecz biorąc pożądane jest równomierne rozłożenie zadań ze względu
na trudność oraz niewystępowanie zadań o niskiej mocy różnicującej.
\bold{Wybór wskaźnika mocy różnicującej}
\itemize{
  \item{\code{Pearson} - korelacja liniowa Pearsona (punktowo-dwuseryjna);
        dla zadań z wąskimi skalami oceny (w szczególności ocenianych
        binarnie: 0 lub 1) w praktyce niemożliwe jest osiągnięcie korelacji
        bliskich jedności; maksymalna możliwa do osiągnięcia korelacja jest
        tym mniejsza, im bardziej trudność/łatwość zadania odbiega od 0.5.}
  \item{\code{dwuseryjna} - korelacja dwuseryjna; adekwatna tylko do zadań
        ocenianych binarnie (0 lub 1), jest zawsze większa od korelacji
        Pearsona, ale jeśli rozkład sumy wyraźnie różni się od rozkładu
        normalnego może przeszacowywać natężenie związku (w tym przyjmować
        bezsensowne wartości większe niż 1);}
  \item{\code{bez zadania} - korelacja liniowa Pearsona (j.w.), ale z sumą
        punktów pomniejszoną o liczbę punktów uzyskanych za dane zadanie;
        bardziej adekwatna od zwykłej korelacji Pearsona dla krótkich testów
        lub w sytuacji, gdy dane zadanie ma dużo szerszą skalę oceny niż
        pozostałe;}
}
\bold{Kolory punktów}
Punkty reprezentujące zadania kolorowane są w zależności od wartości
wskaźnika rzetelności bez zadania (używany jest współczynnik alfa Feldt-Raju,
p. \code{\link{alfaFR}}. Zielone reprezentują zadania, po usunięciu których
reztelność testu spada, a czerwone te, po usunięciu których rzetelność testu
rośnie (ta druga sytuacja jest niepożądana).
\bold{Obwiednia maksymalnych wartości korelacji}
Jeśli argument \code{wsp = 'Pearson'}, na wykresie niebieskim kolorem
naniesiona zostanie obwiednia wyznaczające maksymalne możliwe do uzyskania
wartości korelacji \bold{pomiędzy zadaniem punktowanym binarnie, a sumą
punktów o rozkładzie zbliżonym do normalnego}. Jeśli pewne zadania mają
szersze skale oceny i/lub rozkład sumy wyraźnie różni się od rozkładu
normalnego, możliwe jest uzyskanie wyższych wartości korelacji.
}
\examples{
wykres_lmr(symTest)
}
\seealso{
\code{\link{parametry_zadan}}
}

