% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wykres_rd.R
\name{wykres_rd}
\alias{wykres_rd}
\title{Rozkład wyboru dystraktorów}
\usage{
wykres_rd(x, poprawnaOdpowiedz, odpowiedzi = NULL, verbose = TRUE)
}
\arguments{
\item{x}{wektor odpowiedzi na zadanie}

\item{poprawnaOdpowiedz}{ciąg znaków lub liczba będąca poprawną odpowiedzią}

\item{odpowiedzi}{opcjonalnie wektor tekstowy lub liczbowy zawierający
wszystkie wartości, jakie może przyjąć wynik zadania}

\item{verbose}{wartość logiczna - czy wydrukować wyniki analizy}
}
\value{
Funkcja zwraca milcząco wektor z częstościami wyboru dystraktorów.
}
\description{
Funkcja rysuje wykres rozkładu częstości wyboru dystraktorów
w danym zadaniu.
}
\details{
Jako dystraktory rozumie się tu tylko niepoprawne odpowiedzi (odpowiedź
poprawna nie zostanie uwzględniona na wykresie).

Argument \code{odpowiedzi} powinien zostać wykorzystany, jeśli niektóre
dystraktory w ogóle nie zostały wybrane - inaczej nie będą mogły być
uwzględnione na wykresie.
}
\examples{
wykres_rd(daneEWD$Z4, "B")
}

