% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wykres_rs.R
\name{wykres_rs}
\alias{wykres_rs}
\title{Rozkład sumy punktów}
\usage{
wykres_rs(x, maks = NULL, min = NULL, na.rm = TRUE, verbose = TRUE)
}
\arguments{
\item{x}{macierz typu \code{numeric} lub ramka danych (data frame)
zawierająca zmienne typu \code{numeric}}

\item{maks}{opcjonalnie wektor liczb całkowitych opisujący maksymalną
liczbę puntków możliwych do uzyskania za poszczególne zadania}

\item{min}{opcjonalnie wektor liczb całkowitych opisujący minimalną
wartość, jaką może przyjąć wynik poszczególnych zadań}

\item{na.rm}{wartość logiczna - czy przy obliczeniach ignorować braki danych}

\item{verbose}{wartość logiczna - czy wydrukować wyniki analizy}
}
\value{
Funkcja nic nie zwraca.
}
\description{
Funkcja rysuje wykres rozkładu sumy punktów uzyskanych w teście.
}
\details{
Szczegóły związane z użyciem argumentów \code{maks} i \code{min} opisane są
w sekcji \code{Details} pomocy do funkcji \code{\link{latwosc}}
i \code{\link{trudnosc}}.
}
\examples{
wykres_rs(wynikiSymTest)
}
\seealso{
\code{\link{normy_staninowe}}, \code{\link{alfaC}}
}

