% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wykres_rwz.R
\name{wykres_rwz}
\alias{wykres_rwz}
\title{Rozkład wyników zadania}
\usage{
wykres_rwz(x, wartosci = NULL, verbose = TRUE)
}
\arguments{
\item{x}{macierz typu \code{numeric} lub ramka danych (data frame)
zawierająca zmienne typu \code{numeric}}

\item{wartosci}{opcjonalnie wektor liczb zawierający wszystkie wartości,
jakie może przyjąć wynik zadania}

\item{verbose}{wartość logiczna - czy wydrukować wyniki analizy}
}
\value{
Funkcja zwraca ramkę danych z rozkładem punktów uzyskanych za zadanie.
}
\description{
Funkcja rysuje wykres rozkładu sumy punktów uzyskanych w teście.
}
\examples{
wykres_rwz(apply(wynikiSymTest[, 1:4], 1, sum))
}

