% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wykres_twg.R
\name{wykres_twg}
\alias{wykres_twg}
\title{Rozkład trudności zadania w podziale na grupy}
\usage{
wykres_twg(x, grupy, nazwaGrupowania = "grupa", oznaczMonotonicznosc = TRUE,
  maks = NA, min = NA, na.rm = TRUE, verbose = TRUE)
}
\arguments{
\item{x}{wektor liczb zawierający wyniki zadania}

\item{grupy}{wektor opisujący przynależność obserwacji do grup}

\item{nazwaGrupowania}{ciąg znaków - nazwa, która zostanie wyświetlona jako
etykieta osi poziomej na wykresie}

\item{oznaczMonotonicznosc}{wartość logiczna - czy ma być analizowana
monotoniczność trudności ze względu na podane grupowanie}

\item{maks}{opcjonalnie wektor liczb całkowitych opisujący maksymalną
liczbę puntków możliwych do uzyskania za poszczególne zadania}

\item{min}{opcjonalnie wektor liczb całkowitych opisujący minimalną
wartość, jaką może przyjąć wynik poszczególnych zadań}

\item{na.rm}{wartość logiczna - czy przy obliczeniach ignorować braki danych}

\item{verbose}{wartość logiczna - czy wydrukować wyniki analizy}
}
\value{
Funkcja zwraca milcząco wektor z oszacowaniami trudności w ramach grup.
}
\description{
Funkcja rysuje wykres trudności zadania w ramach podanych grup.
}
\details{
\bold{Grupowanie}

Na potrzeby rysowania wykresu grupy zostaną posorotowane. Jeśli chce się
uzyskać kolejność prezentacji inną, niż wynikającą z posortowania unikalnych
wartości wektora podanego jako argument \code{grupy}, należy wcześniej
przerobić go na \link[base]{factor} zadając pożądaną kolejność jego poziomów.

Jeśli argument \code{oznaczMonotonicznosc} przyjmuje wartość \code{TRUE},
przeanalizowana zostanie monotoniczność trudności ze względu na podział na
grupy. Dokładniej, oznaczone zostaną te grupy, w których trudność jest
większa, niż w poprzedniej. Dla pewnych pogrupowań, np. ze względu na wyniki
całego testu (w szczególności wyrażone jako stanin) ma to sens - zaburzenia
monotoniczności świadczą wtedy o problemie z zadaniem. Dla innych podziałów,
np. ze względu na płeć, typowo nie ma to sensu i wtedy należy wywołać funkcję
z argumentem \code{oznaczMonotonicznosc} ustawionym na \code{FALSE}.

\bold{Maksymalne i minimalne wartości}

Domyślnie, zgodnie z konwencją typową dla testów wiedzy/umiejętności,
jako minimalna możliwa do przyjęcia wartość wyniku zadania przyjmowane jest 0.
Maksymalna możliwa do przyjęcia wartość określana jest z kolei empirycznie,
jako największa wartość, która występuje w danych.

Wykorzystanie argumentu \code{maks} jest konieczne, gdy ze względu na
niewielką liczbę obserwacji w zbiorze i/lub dużą trudność zadania dla
pewnych zadań w danych nie występuje maksymalna możliwa do osiągnięca liczba
punktów. Wartość przekazywana argumentem \code{maks} może być brakiem
danych - maksymalna możliwa do uzyskania wartość zostanie wtedy określona na
podstawie danych (j.w.).

Wykorzystanie argumentu \code{min} jest konieczne w przypadku niektórych
schematów kodowania odpowiedzi na pytania np. w testach psychologicznych,
używających skal Likerta, czy dyferencjałów semantycznych. Często w takich
przypadkach odpowiedzi kodowane są kolejnymi liczbami naturalnymi, począwszy
od 1. Ogólnie rzecz biorąc, jeśli minimalną wartość, jaką może przyjąć wynik
danego zadania jest różna od 0, należy podać ją przy pomocy argumentu
\code{min}. Wartość przekazywana argumentem \code{min} może być brakiem
danych - jako minimalna możliwa do uzyskania wartość zostanie wtedy przyjęte 0.

\bold{Obliczania łatwości/trudności dla zadań o minimalnej możliwej wartości różnej od 0}

Jeśli minimalna możliwa do uzyskania wartość wyniku zadania jest różna od 0,
łatwość obliczona zostanie jako \code{E(X - min) / (max - min)}, gdzie
\code{X} oznacza wynik zadania, a \code{min} i \code{max} odpowiednio
najmniejszą i największą możliwą do uzyskania wartość wyniku zadania.
}
\examples{
grupy = normy_staninowe(wynikiSymTest)$wynikiStaninowe
wykres_twg(wynikiSymTest[, 1], grupy)
wykres_twg(wynikiSymTest[, 1], grupuj_staniny(grupy))
}
\seealso{
\code{\link{trudnosc}}
}

