% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wynikiSymTest.R
\docType{data}
\name{wynikiSymTest}
\alias{wynikiSymTest}
\title{Symulowane wyniki testu zlozonego z 10 zadan}
\format{data frame: 500 obserwacje, 10 zmiennych}
\description{
Symulowane wyniki testu złożonego z 10 zadań.
}
\keyword{datasets}

