% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wyrownaj_do_lewej.R
\name{wyrownaj_do_lewej}
\alias{wyrownaj_do_lewej}
\title{Formatowanie liczb}
\usage{
wyrownaj_do_lewej(x)
}
\arguments{
\item{x}{wektor tekstowy}
}
\value{
wektor tekstowy
}
\description{
Funkcja wyrównuje podane ciągi znaków na miejscu dziesiętnym.
}

