% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trudnosc.R
\name{latwosc}
\alias{latwosc}
\title{Łatwość zadań}
\usage{
latwosc(x, maks = NULL, na.rm = TRUE, verbose = TRUE)
}
\arguments{
\item{x}{macierz typu \code{numeric} lub ramka danych (data frame)
zawierająca zmienne typu \code{numeric}}

\item{maks}{opcjonalnie wektor liczb całkowitych opisujący maksymalną
liczbę puntków możliwych do uzyskania za poszczególne zadania}

\item{na.rm}{wartość logiczna - czy przy obliczeniach ignorować braki danych}

\item{verbose}{wartość logiczna - czy wydrukować wyniki analizy}
}
\value{
Funkcja zwraca milcząco wektor z oszacowaniami łatwości zadań.
}
\description{
Funkcja służy do oszacowania łatwości zadań.
}
\examples{
latwosc(symTest)
}

