% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moc_roznicujaca.R
\name{moc_roznicujaca}
\alias{moc_roznicujaca}
\title{Moc różnicująca}
\usage{
moc_roznicujaca(x, na.rm = TRUE, verbose = TRUE)
}
\arguments{
\item{x}{macierz typu \code{numeric} lub ramka danych (data frame)
zawierająca zmienne typu \code{numeric}}

\item{na.rm}{wartość logiczna - czy przy obliczeniach ignorować braki danych}

\item{verbose}{wartość logiczna - czy wydrukować wyniki analizy}
}
\value{
Funkcja zwraca milcząco dwuelementową listę, której elementy zawierają:
\itemize{
   \item{\code{korelacjePearsona} wartości współczynników korelacji Pearsona
         (punktowo-dwuseryjnych) pomiędzy wynikiem całego testu, a wynikami
         poszczególnych zadań,}
   \item{\code{korelacjeDwuseryjne} wartości współczynników korelacji
         dwuseryjnych pomiędzy wynikiem całego testu, a wynikami
         poszczególnych zadań.}
}
}
\description{
Funkcja służy do oszacowania mocy różnicującej zadań.
}
\details{
\itemize{
  \item{\code{korelacja liniowa Pearsona (punktowo-duseryjna)}:
        dla zadań z wąskimi skalami oceny (w szczególności ocenianych
        binarnie: 0 lub 1) w praktyce niemożliwe jest osiągnięcie korelacji
        bliskich jedności; maksymalna możliwa do osiągnięcia korelacja jest
        tym mniejsza, im bardziej trudność/łatwość zadania odbiega od 0.5.}
  \item{\code{korelacja dwuseryjna}: adekwatna tylko do zadań ocenianych
        binarnie (0 lub 1) - dla innych nie jest tu obliczana; jest zawsze
        większa (co do wartości bezwzględnej) od korelacji Pearsona, ale
        jeśli rozkład sumy punktów wyraźnie różni się od rozkładu normalnego,
        może przeszacowywać natężenie związku (w tym przyjmować bezsensowne
        wartości większe niż 1);}
  \item{\code{korelacja Pearsona bez zadania} - korelacja liniowa Pearsona
        (j.w.), ale z sumą punktów pomniejszoną o liczbę punktów uzyskanych
        za dane zadanie; bardziej adekwatna od zwykłej korelacji Pearsona dla
        krótkich testów lub w sytuacji, gdy dane zadanie ma dużo szerszą
        skalę oceny niż pozostałe;}
}
}
\examples{
moc_roznicujaca(symTest)
}
\seealso{
\code{\link{parametry_zadan}}, \code{\link{wykres_lmr}}
}

