% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normy_staninowe.R
\name{normy_staninowe}
\alias{normy_staninowe}
\title{Normy staninowe dla sumy punktow}
\usage{
normy_staninowe(x, maks = NULL, min = NULL, na.rm = TRUE,
  verbose = TRUE)
}
\arguments{
\item{x}{macierz typu \code{numeric} lub ramka danych (data frame)
zawierająca zmienne typu \code{numeric}}

\item{maks}{opcjonalnie wektor liczb całkowitych opisujący maksymalną
liczbę puntków możliwych do uzyskania za poszczególne zadania}

\item{min}{opcjonalnie wektor liczb całkowitych opisujący minimalną
wartość, jaką może przyjąć wynik poszczególnych zadań}

\item{na.rm}{wartość logiczna - czy przy obliczeniach ignorować braki danych}

\item{verbose}{wartość logiczna - czy wydrukować wyniki analizy}
}
\value{
Funkcja zwraca milcząco dwuelementową listę, której elementy zawierają:
\itemize{
   \item{\code{normyStaninowe} ramka danych zawierająca 9 obserwacji, dla
         opisywane przez kolumny: \code{stanin}, \code{min_pkt},
         \code{maks_pkt};}
   \item{\code{wynikiStaninowe} wektor liczbowy o długości równej liczbie
         wierszy w \code{x}, zawierający wyniki poszczególnych zdających
         wyrażone na skali staninowej.}
}
}
\description{
Funkcja oblicza normy staninowe dla sumy punktów i przelicza
wyniki zdających na skalę staninową.
}
\examples{
normy_staninowe(wynikiSymTest)
}

