% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wykres_rwz.R
\name{wykres_rwz}
\alias{wykres_rwz}
\title{Rozklad wynikow zadania}
\usage{
wykres_rwz(x, wartosci = NULL, verbose = TRUE)
}
\arguments{
\item{x}{wektor liczb zawierający wyniki zadania}

\item{wartosci}{opcjonalnie wektor liczb zawierający wszystkie wartości,
jakie może przyjąć wynik zadania}

\item{verbose}{wartość logiczna - czy wydrukować wyniki analizy}
}
\value{
Funkcja zwraca ramkę danych z rozkładem punktów uzyskanych za zadanie.
}
\description{
Funkcja rysuje wykres rozkładu sumy punktów uzyskanych w zadaniu.
}
\examples{
wykres_rwz(apply(wynikiSymTest[, 1:4], 1, sum))
}
\seealso{
\code{\link{parametry_zadan}}
}

